Application/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.3; python_version >= '3.6'
cx-oracle==8.3.0
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask==2.0.2
greenlet==1.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
itsdangerous==2.0.1; python_version >= '3.6'
jaydebeapi==1.2.3
jinja2==3.0.3; python_version >= '3.6'
jpype1==1.3.0; python_version > '2.7' and platform_python_implementation != 'Jython'
markupsafe==2.0.1; python_version >= '3.6'
python-dotenv==0.19.2
sqlalchemy==1.4.27
werkzeug==2.0.2; python_version >= '3.6'
